requirements.txt
fastapi[full]
uvicorn
pandas
databases
mysql-connector-python
httpx
redis
aiomysql
apscheduler
pydantic-sqlalchemy
schedule